Migration classes will first check if changes need to be integrated before executing them.

+++ src/application/migrations/007_add_service_availabilities_type.php
@@ -14,25 +14,26 @@
 class Migration_Add_service_availabilities_type extends CI_Migration {
     public function up()
     {
-        $this->load->dbforge();
-
-        $fields = [
-            'availabilities_type' => [
-                'type' => 'VARCHAR',
-                'constraint' => '32',
-                'default' => 'flexible',
-                'after' => 'description'
-            ]
-        ];
-
-        $this->dbforge->add_column('ea_services', $fields);
-
-        $this->db->update('ea_services', ['availabilities_type' => 'flexible']);
+        if (!$this->db->field_exists('availabilities_type', 'ea_services')) {
+            $fields = [
+                'availabilities_type' => [
+                    'type' => 'VARCHAR',
+                    'constraint' => '32',
+                    'default' => 'flexible',
+                    'after' => 'description'
+                ]
+            ];
+    
+            $this->dbforge->add_column('ea_services', $fields);
+    
+            $this->db->update('ea_services', ['availabilities_type' => 'flexible']);
+        }
     }
 
     public function down()
     {
-        $this->load->dbforge();
-        $this->dbforge->drop_column('ea_services', 'availabilities_type');
+        if ($this->db->field_exists('availabilities_type', 'ea_services')) {
+            $this->dbforge->drop_column('ea_services', 'availabilities_type');
+        }
     }
 }
