+++ src/application/migrations/008_add_service_attendants_number.php
@@ -14,25 +14,26 @@
 class Migration_Add_service_attendants_number extends CI_Migration {
     public function up()
     {
-        $this->load->dbforge();
-
-        $fields = [
-            'attendants_number' => [
-                'type' => 'INT',
-                'constraint' => '11',
-                'default' => '1',
-                'after' => 'availabilities_type'
-            ]
-        ];
-
-        $this->dbforge->add_column('ea_services', $fields);
-
-        $this->db->update('ea_services', ['attendants_number' => '1']);
+        if (!$this->db->field_exists('attendants_number', 'ea_services')) {
+            $fields = [
+                'attendants_number' => [
+                    'type' => 'INT',
+                    'constraint' => '11',
+                    'default' => '1',
+                    'after' => 'availabilities_type'
+                ]
+            ];
+    
+            $this->dbforge->add_column('ea_services', $fields);
+    
+            $this->db->update('ea_services', ['attendants_number' => '1']);
+        }
     }
 
     public function down()
     {
-        $this->load->dbforge();
-        $this->dbforge->drop_column('ea_services', 'attendants_number');
+        if (!$this->db->field_exists('attendants_number', 'ea_services')) {
+            $this->dbforge->drop_column('ea_services', 'attendants_number');
+        }
     }
 }
